Move slackbot logic to another class
